refractorized the assessment
refractorized the evaluation so as not to have duplicate code, inserted in utilities
